backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-rest-knox==5.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.11.0
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3